Change test config

+++ database/migrations/2021_03_24_161000_refactor_without_sections_and_content_urls.php
@@ -22,9 +22,11 @@ class RefactorWithoutSectionsAndContentUrls extends Migration
         
         $this->adaptPagesTable();
 
-        Schema::dropIfExists("sections");
-        Schema::dropIfExists("pagegroups");
-        Schema::dropIfExists("content_urls");
+        if(config("database.default") != "sqlite") {
+            Schema::dropIfExists("sections");
+            Schema::dropIfExists("content_urls");
+            Schema::dropIfExists("pagegroups");
+        }
     }
 
     private function migrateSectionsToPages()
